MAINT: bump neovim version
- Rolling in my own version of gruvbox-material

This is just a fork of `gruvbox-baby` that supports the material colors
from `gruvbox-material`. Reason for this is just because I am lazy and
wanted to create a theme that supports the plugins I am using, because I
cannot be bothered implementing myself, I just stole a preconfigured one
and went hard with grep to change things over.

Post-grep, I didn't like some of the colors, so switched them around a
bit to get something similar to what I had previously.