cmp,ai: migrating to blink.cmp, introducing supermaven-nvim
Patch reworks the cmp and ai integrations to use the new blink.cmp
plugin. This also introduces supermaven-nvim as an optional dependency.

By default the ai integration is disabled, but can be enabled by setting
the global variable `vim.g.ai_integration_enabled = true` in settings.lua.

The intention on ongoing patches is to refactor some of the code to make
these options set at the top level of the config. Eventually it would be
good to write a small plugin or have it internal to the config to parse
a project-specific file for setting up flags in the vim.global namespace.

> Motivation: Some work projects are not allowed to use external AI
integrations, supporting both is optimal, following the same system that
is used to remove dependencies for the docker instance of this config.