project-settings: Enable setting `vim.g` per-project to enable/disable features
Introduces internal-plugin `project-settings` into neovim, this plugin
can read a per-project setup file to configure AI completion, nerd fonts
and in future other flags that we don't want on all the time.

These globals are also configured to read from the environment
variables, this eliminates the hack that the `jordanp-env` has to sed in
changes to settings files, and instead can just supply env config to
override this (this env doesn't have nerd fonts for example).